python-scraper/requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
requests>=2.28.0
beautifulsoup4>=4.11.0
selenium>=4.10.0
webdriver-manager>=3.8.0
lxml>=4.9.0
fake-useragent>=1.2.0
gunicorn>=20.0.0
openai>=1.0.0
python-dotenv>=1.0.0